HCC Program/requirements.txt
alembic==0.8.4
appnope==0.1.0
blinker==1.4
cycler==0.10.0
decorator==4.0.9
dmidecode==0.8.1
dominate==2.1.17
eventlet==0.18.4
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Moment==0.5.1
Flask-MySQLdb==0.2.0
Flask-Script==2.0.5
Flask-SocketIO==2.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
get-0.0.4==0.0.4
gnureadline==6.3.3
greenlet==0.4.9
ipython==4.1.2
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
matplotlib==1.5.1
mysqlclient==1.3.7
numpy==1.11.0
pandas==0.18.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psutil==4.1.0
ptyprocess==0.5.1
pyparsing==2.1.1
python-csv==0.0.9
python-dateutil==2.5.2
python-editor==0.5
python-engineio==0.9.0
python-socketio==1.2
pytz==2016.3
request==0.0.2
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.12
traitlets==4.1.0
visitor==0.1.2
Werkzeug==0.11.4
WTForms==2.1
xlrd==0.9.4
